==> 1_assign_treatments.Rmd <==
---
title: "Assigning Treatments"
author: "Brandi Pessman"
date: "2024-10-15"
output: html_document
---

# Load Libraries

```{r libraries}
library(tidyverse) # for wrangling data
```

# Import Data

```{r import}
set.seed(2023)

data <- read.csv("data/spiders_20230905.csv") %>% 
  select(-Sex)

urban <- data %>% 
  filter(Site == "Urban")
rownames(urban) <- urban$ID
urban <- select(urban, -ID)

rural <- data %>% 
  filter(Site == "Rural")
rownames(rural) <- rural$ID
rural <- select(rural, -ID)

UrbanID = list()
UrbanAge = list()
RuralID = list()
RuralAge = list()
remaining = list()
```

# Match Urban and Rural by Age

```{r match by age}
for (spider in rownames(urban)) {
  UrbanID <- append(UrbanID, spider)
  UrbanAge <- append(UrbanAge, urban[spider, "CurrentAge"])
  age = rural[rural["CurrentAge"] == min(rural$CurrentAge), ]
  while (length(UrbanID) != length(RuralID)) {
    potential_match = age[sample(nrow(age), 1), ]
    if (abs(potential_match$CurrentAge - urban[spider, "CurrentAge"]) <= 6) {
      rural = rural[!rownames(rural) %in% rownames(potential_match), ]
      RuralID <- append(RuralID, rownames(potential_match))
      RuralAge <- append(RuralAge, potential_match$CurrentAge)
    }
    else {
      remaining <- append(remaining, rownames(age))
      rural = rural[!rownames(rural) %in% rownames(age), ]
      age = rural[rural["CurrentAge"] == min(rural$CurrentAge), ]
    }
  }
}
```

# Spiders and ages not assigned

```{r remaining}
urban <- data %>% 
  filter(Site == "Urban")
rownames(urban) <- urban$ID
urban <- select(urban, -ID)

rural <- data %>% 
  filter(Site == "Rural")
rownames(rural) <- rural$ID
rural <- select(rural, -ID)

remaining_df = data.frame(rural[rownames(rural) %in% remaining, ])
```

# Data formatting and assignment of iterations, treatments, and bins (oldest go first)

```{r assign}
treatments = data.frame(cbind(UrbanID, UrbanAge, RuralID, RuralAge))
treatments = treatments[rev(order(as.integer(treatments$UrbanAge))), ]
treatments$iterations <- rep(seq(1, 10), each = 3, length.out = 30)
treatments$treatment <- rep(c(sample(c("Loud", "Quiet", "Loud", "Quiet", "Loud", "Quiet"))), length.out = 30)
urban_asm <- treatments %>% 
  select(UrbanID, UrbanAge, treatment, iterations) %>% 
  mutate(Site = "Urban",
         ID = UrbanID, 
         Age = UrbanAge) %>% 
  select(-UrbanID, -UrbanAge)
rural_asm <- treatments %>% 
  select(RuralID, RuralAge, treatment, iterations) %>% 
  mutate(Site = "Rural",
         ID = RuralID, 
         Age = RuralAge) %>% 
  select(-RuralID, -RuralAge)
treatments = rbind(urban_asm, rural_asm)
treatments <- treatments %>% 
  select(ID, Site, Age, iterations, treatment) %>% 
  mutate(Age = as.integer(Age))
treatments = treatments[order(treatments$iterations), ]
# to assign bins
sample(1:6, 6)
treatments <- treatments[order(treatments$iterations),] 
```

# Check for normal distributioin

```{r normal}
ul <- treatments %>% 
  filter(Site == "Urban",
         treatment == "Loud")
shapiro.test(ul$Age)
uq <- treatments %>% 
  filter(Site == "Urban",
         treatment == "Quiet")
shapiro.test(uq$Age)
rl <- treatments %>% 
  filter(Site == "Rural",
         treatment == "Loud")
shapiro.test(rl$Age)
rq <- treatments %>% 
  filter(Site == "Rural",
         treatment == "Quiet")
shapiro.test(rq$Age)

treatments %>% 
    group_by(Site, treatment) %>% 
    count()
```

# Test that age doesn't vary by origin and treatment

```{r test}
stats <- summary(lm(Age ~ Site * treatment, data = treatments))
stats
```

# Save dataframe

```{r save}
# if age doesn't vary, save the dataframe
if (stats$coefficients[2, 4] > 0.05 | stats$coefficients[3, 4] > 0.05 | stats$coefficients[4, 4] > 0.05){
  treatments <- apply(treatments, 2, as.character)
  write.csv(treatments, "wrangled_data/treatments_assigned.csv", row.names = FALSE)
}
```
