==> 4_vibration_transmission.Rmd <==
---
title: "Web Vibration Transmission Analysis"
author: "Brandi Pessman"
date: "2024-11-19"
output: html_document
---

# Load Libraries

```{r libraries}
library(tidyverse) # for wrangling data
library(MASS) # glm.nb
library(mgcv) # gam
library(broom) # augment
library(ggpubr) # ggarrange
library(emmeans) # emmeans
library(flextable) # flextable
```

# Load Other 

```{r load other}
options(scipen = 999)
size = 24
my_theme <- theme_classic() +
  theme(axis.text = element_text(size = size, color = "black", family = "sans"),
        axis.title = element_text(size = size, color = "black", family = "sans"),
        legend.text = element_text(size = size, color = "black", family = "sans"),
        legend.title = element_text(size = size, color = "black", family = "sans"),
        strip.text = element_text(size = size, color = "black", family="sans"),
        plot.title = element_text(size = size, color = "black", family="sans"),
        legend.position = "top")

two_colors = c("#1b9e77", "#d95f02")
four_colors = c("black", "#1b9e77", "grey70", "#d95f02")
#four_colors = c("#90ecd1", "#1b9e77", "#febd8c", "#d95f02")

treatment.labs <- c("Quiet (4 Days)", "Loud (4 Days)")
names(treatment.labs) <- c("Quiet", "Loud")
```

# Load Data

```{r load data}
playback <- readRDS("wrangled_data/playback.rds")
example <- readRDS("wrangled_data/example.rds")
stimulus <- readRDS("wrangled_data/stimulus.rds")
vibration_transmission <- readRDS(file = "wrangled_data/vibration_transmission.rds")
vibration_transmission2 <- readRDS(file = "wrangled_data/vibration_transmission2.rds")
spiders <- vibration_transmission %>% 
  dplyr::select(SpiderID, Site, treatment, Site_Treatment, trial_age, mass_mg, condition, wet_mg, dry_mg, moisture_per) %>% 
  unique()
silk <- spiders %>% 
  mutate(dry_microg_body_mg = (dry_mg * 1000) / mass_mg,
         wet_microg_body_mg = (wet_mg * 1000) / mass_mg,
         moisture_per = as.numeric(moisture_per) / 100,
         treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud")) %>% 
  filter(! dry_mg == 0) 
```

# Descriptives

## Age and Condition

```{r age}
spiders %>% 
  group_by(Site, treatment) %>% 
  summarize(mean_age = mean(trial_age),
            se_age = plotrix::std.error(trial_age)) %>% 
  ggplot() +
  geom_point(aes(x = Site, y = mean_age, color = treatment)) +
  geom_errorbar(aes(x = Site, ymax = mean_age + se_age, ymin = mean_age - se_age, color = treatment)) +
  my_theme

# try a few differnt models but all give the same results
shapiro.test(rank(spiders$trial_age))
car::Anova(lm(rank(trial_age) ~ Site_Treatment, data = spiders))
car::Anova(glm.nb(trial_age ~ Site_Treatment, data = spiders))
car::Anova(lm(trial_age ~ Site_Treatment, data = spiders)) # reported this test
```

```{r condition}
spiders %>% 
  group_by(Site, treatment) %>% 
  summarize(mean_cond = mean(condition),
            se_cond = plotrix::std.error(condition)) %>% 
  ggplot() +
  geom_point(aes(x = Site, y = mean_cond, color = treatment)) +
  geom_errorbar(aes(x = Site, ymax = mean_cond + se_cond, ymin = mean_cond - se_cond, color = treatment)) +
  my_theme

shapiro.test(spiders$condition)
car::Anova(lm(condition ~ Site_Treatment, data = spiders))
```

```{r mass}
spiders %>% 
  group_by(Site, treatment) %>% 
  summarize(mean_mass = mean(mass_mg),
            se_mass = plotrix::std.error(mass_mg)) %>% 
  ggplot() +
  geom_point(aes(x = Site, y = mean_mass, color = treatment)) +
  geom_errorbar(aes(x = Site, ymax = mean_mass + se_mass, ymin = mean_mass - se_mass, color = treatment)) +
  my_theme

shapiro.test(spiders$mass_mg)
car::Anova(lm(mass_mg ~ Site_Treatment, data = spiders))
```

## Playback Conditions During Web Construction

```{r quiet vs loud amplitude}
playback %>% 
  filter(Low_Freq < 1000) %>% 
  group_by(Treatment) %>% 
  summarize(mean(snr_db),
            plotrix::std.error(snr_db))
```

```{r playback loess curves}
# loess curves
playback_quiet <- playback %>% 
  filter(Treatment == "Quiet")
loess_pb_quiet <- loess(snr_db ~ Low_Freq, data = playback_quiet, span = 0.5)
nd <- expand_grid(Low_Freq = seq(21.5, 1996, 5.38),
                  Treatment = "Quiet") 
predictions_pb_quiet <- predict(loess_pb_quiet, newdata = nd, se = TRUE)
predictions_pb_quiet <- cbind(nd, predictions_pb_quiet)
playback_loud <- playback %>% 
  filter(Treatment == "Loud")
loess_pb_quiet <- loess(snr_db ~ Low_Freq, data = playback_loud, span = 0.5)
nd <- expand_grid(Low_Freq = seq(21.5, 1996, 5.38),
                  Treatment = "Loud") 
predictions_pb_loud <- predict(loess_pb_quiet, newdata = nd, se = TRUE)
predictions_pb_loud <- cbind(nd, predictions_pb_loud)
predictions_pb <- rbind(predictions_pb_quiet, predictions_pb_loud) %>% 
  mutate(Treatment = fct_relevel(Treatment, "Quiet", "Loud"))

ggplot() +
  geom_vline(xintercept = 1000, linetype = "dashed", color = "black") +
  geom_line(aes(x = Low_Freq, y = snr_db, color = Treatment), linewidth = 0.5, alpha = 0.15, data = playback) +
  geom_line(aes(x = Low_Freq, y = fit, color = Treatment), data = predictions_pb) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - se.fit, ymax = fit + se.fit, fill = Treatment), alpha = 0.5, data = predictions_pb) +
    xlab("Frequency (Hz)") +
  ylab("Amplitude (Signal-to-Noise Ratio, dB)\nQuieter <------------> Louder") +
  scale_color_manual("Treatment", values = two_colors) +
  scale_fill_manual("Treatment", values = two_colors) +
  scale_x_continuous(limits = c(20, 2000), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, 65), breaks = c(seq(0, 60, 10)), expand = c(0,0)) +
  my_theme
```

```{r playback gam model}
# generalized additive model
gam_model <- gam(snr_db ~  Treatment + s(Low_Freq, by = Treatment, k = 10), data = playback) #better
gam_model2 <- gam(snr_db ~  s(Low_Freq, k = 10), data = playback)
results <- anova(gam_model2, gam_model, test = "Chisq")
c(AIC(gam_model), AIC(gam_model2))
sum1 <- summary(gam_model)
sum2 <- summary(gam_model2)
sum1
sum2
sum1$pTerms.table
gam.check(gam_model)

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Treatment = levels(factor(playback$Treatment))))
predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions <- cbind(nd, predictions) %>% 
  mutate(Treatment = fct_relevel(Treatment, "Quiet", "Loud"))

size = 12
ggplot() +
  geom_vline(xintercept = 1000, color = "black", linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = snr_db, color = Treatment), alpha = 0.15, data = playback) +
  geom_line(aes(x = Low_Freq, y = fit, color = Treatment), data = predictions) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Treatment), alpha = 0.5, data = predictions) +
  scale_y_continuous(limits = c(-5, 65), breaks = c(seq(0,60,10)), expand = c(0,0)) + 
  scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  ylab("Signal-to-Noise Ratio (dB)\nQuieter <----------------------> Louder") +
  #ylab("Amplitude (dB)\nQuieter <----------------------> Louder") +
  scale_fill_manual("", values = two_colors) +
  scale_color_manual("", values = two_colors) +
  my_theme
# 700 x 700
```

# Example Calculations (Method 1)

## Signal-to-Noise Ratio of Stimulus

```{r stimulus raw}
loess_ambient <- loess(Power_ambient_stim ~ Low_Freq, data = stimulus, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Ambient")
predictions_ambient <- predict(loess_ambient, newdata = nd, se = TRUE)
predictions_ambient <- cbind(nd, predictions_ambient)

loess_test <- loess(Power_test_stim ~ Low_Freq, data = stimulus, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Test")
predictions_test <- predict(loess_test, newdata = nd, se = TRUE)
predictions_test <- cbind(nd, predictions_test)

predictions_ambtest <- rbind(predictions_ambient, predictions_test) 

size = 16
ggplot() +
  geom_line(aes(x = Low_Freq, y = Power_ambient_stim), linetype = "dashed", color = "grey", data = stimulus, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = Power_test_stim), linetype = "dashed", color = "gold", data = stimulus, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit, color = Type), linetype = "dashed", linewidth = 1, data = predictions_ambtest) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Type), alpha = 0.5, data = predictions_ambtest) +
  xlab("Frequency (Hz)") +
  ylab("Inband Power (dB re FS)") +
  #ylab("Amplitude (dB)") +
  scale_color_manual("",
                       values = c("grey", "gold"),
                       labels = c("Ambient",
                                 "Test")) +
  scale_fill_manual("",
                       values = c("grey", "gold"),
                       labels = c("Ambient",
                                 "Test")) +
  scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-120, 2), expand = c(0, 0)) +
  guides(linetype = guide_legend(override.aes = list(linewidth = 0.5) ) ) +
  my_theme +
  theme(legend.position = "none")
# 800 x 600
```

```{r stimulus snr}
# stimulus of attenuation with loess
loess_snr <- loess(snr_db_stim ~ Low_Freq, data = stimulus, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5))
predictions_snr <- predict(loess_snr, newdata = nd, se = TRUE)
predictions_snr <- cbind(nd, predictions_snr)

ggplot() +
  geom_hline(yintercept = 0, linetype = "dashed", linewidth = 1, color = "grey") +
  geom_line(aes(x = Low_Freq, y = snr_db_stim), linetype = "dashed", color = "gold", data = stimulus, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit), linetype = "dashed", color = "gold", linewidth = 1, data = predictions_snr) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - se.fit, ymax = fit + se.fit), fill = "gold", alpha = 0.5, data = predictions_snr) +
  xlab("Frequency (Hz)") +
  ylab("Signal-to-Noise Ratio (dB)") +
  scale_x_continuous(limits = c(20, 2100), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-2, 80), breaks = c(seq(0, 70, 10)), expand = c(0,0)) +
  guides(linetype = guide_legend(override.aes = list(linewidth = 0.5) ) ) +
  my_theme
```

## Signal-to-Noise Ratio of Example

```{r example raw}
loess_ambient <- loess(Power_ambient_trial ~ Low_Freq, data = example, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Ambient")
predictions_ambient <- predict(loess_ambient, newdata = nd, se = TRUE)
predictions_ambient <- cbind(nd, predictions_ambient)

loess_test <- loess(Power_test_trial ~ Low_Freq, data = example, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Test")
predictions_test <- predict(loess_test, newdata = nd, se = TRUE)
predictions_test <- cbind(nd, predictions_test)

predictions_ambtest <- rbind(predictions_ambient, predictions_test) 

ggplot() +
  geom_line(aes(x = Low_Freq, y = Power_ambient_trial), linetype = "dashed", color = "grey", data = example, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = Power_test_trial), linetype = "dashed", color = "purple", data = example, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit, color = Type), linetype = "dashed", linewidth = 1, data = predictions_ambtest) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Type), alpha = 0.5, data = predictions_ambtest) +
  xlab("Frequency (Hz)") +
  ylab("Inband Power (dB re FS)") +
  #ylab("Amplitude (dB)") +
  scale_color_manual("",
                       values = c("grey", "purple"),
                       labels = c("Ambient",
                                 "Test")) +
  scale_fill_manual("",
                       values = c("grey", "purple"),
                       labels = c("Ambient",
                                 "Test")) +
  scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-120, 2), expand = c(0, 0)) +
  guides(linetype = guide_legend(override.aes = list(linewidth = 0.5) ) ) +
  my_theme +
  theme(legend.position = "none")
# 800 x 500
```

```{r example snr}
# example of attenuation with loess
loess_snr <- loess(snr_db_trial ~ Low_Freq, data = example, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5))
predictions_snr <- predict(loess_snr, newdata = nd, se = TRUE)
predictions_snr <- cbind(nd, predictions_snr)

ggplot() +
  geom_hline(yintercept = 0, linetype = "dashed", linewidth = 1, color = "grey") +
  geom_line(aes(x = Low_Freq, y = snr_db_trial), linetype = "dashed", color = "purple", data = example, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit), linetype = "dashed", linewidth = 1, data = predictions_snr) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - se.fit, ymax = fit + se.fit), fill = "purple", alpha = 0.5, data = predictions_snr) +
  xlab("Frequency (Hz)") +
  ylab("Signal-to-Noise Ratio (dB)") +
  scale_x_continuous(limits = c(20, 2100), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-2, 80), breaks = c(seq(0, 70, 10)), expand = c(0,0)) +
  my_theme
```

## Energy Loss

```{r raw}
loess_test_stim <- loess(Power_test_stim ~ Low_Freq, data = stimulus, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Stimulus")
predictions_test_stim <- predict(loess_test_stim, newdata = nd, se = TRUE)
predictions_test_stim <- cbind(nd, predictions_test_stim)

loess_test_ex <- loess(Power_test_trial ~ Low_Freq, data = example, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "After Web Transmission")
predictions_test_ex <- predict(loess_test_ex, newdata = nd, se = TRUE)
predictions_test_ex <- cbind(nd, predictions_test_ex)

predictions_stimex <- rbind(predictions_test_stim, predictions_test_ex) %>% 
  mutate(Type = factor(Type),
         Type = fct_relevel(Type, "Stimulus", "After Web Transmission"))

ggplot() +
  geom_line(aes(x = Low_Freq, y = Power_test_stim), color = "gold", data = stimulus, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = Power_test_trial), color = "purple", data = example, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit, color = Type), linewidth = 1, data = predictions_stimex) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Type), alpha = 0.5, data = predictions_stimex) +
  xlab("Frequency (Hz)") +
  ylab("Inband Power (dB re FS)") +
  #ylab("Amplitude (dB)") +
  scale_color_manual("",
                       values = c("gold", "purple"),
                       labels = c("Stimulus", 
                                  "After Web Transmission")) +
  scale_fill_manual("",
                       values = c("gold", "purple"),
                       labels = c("Stimulus", 
                                  "After Web Transmission")) +
  scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-95, -25, 10), expand = c(0, 0)) +
  guides(color = guide_legend(override.aes = list(linewidth = 0.5) ),
         fill = guide_legend(override.aes = list(linewidth = 0.5) )) +
  my_theme
# 500 x 375
```

```{r snr}
loess_stimulus <- loess(snr_db_stim ~ Low_Freq, data = stimulus, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Stimulus")
predictions_stimulus <- predict(loess_stimulus, newdata = nd, se = TRUE)
predictions_stimulus <- cbind(nd, predictions_stimulus)

loess_example <- loess(snr_db_trial ~ Low_Freq, data = example, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Type = "Example")
predictions_example <- predict(loess_example, newdata = nd, se = TRUE)
predictions_example <- cbind(nd, predictions_example)

predictions_stimex <- rbind(predictions_stimulus, predictions_example) 

ggplot() +
  geom_line(aes(x = Low_Freq, y = snr_db_stim), color = "gold", data = stimulus, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = snr_db_trial), color = "purple", data = example, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit, color = Type), linewidth = 1, data = predictions_stimex) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Type), alpha = 0.5, data = predictions_stimex) +
  xlab("Frequency (Hz)") +
  ylab("Signal-to-Noise Ratio (dB)") +
  #ylab("Amplitude (dB)") +
  scale_color_manual("",
                       values = c("purple", "gold"),
                       labels = c("Example",
                                 "Stimulus")) +
  scale_fill_manual("",
                       values = c("purple", "gold"),
                       labels = c("Example",
                                 "Stimulus")) +
  scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, 80, 20), expand = c(0, 0)) +
  guides(color = guide_legend(override.aes = list(linewidth = 0.5) ),
         fill = guide_legend(override.aes = list(linewidth = 0.5) )) +
  my_theme
# 800 x 600
```

```{r energy loss}
# example of attenuation with loess
stimtrial <- left_join(stimulus, example, by = "Low_Freq") %>% 
  dplyr::select(Low_Freq, snr_db_trial, snr_db_stim) %>% 
  mutate(snr_db = snr_db_trial - snr_db_stim) %>% 
  drop_na()

loess_loss <- loess(snr_db ~ Low_Freq, data = stimtrial, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5))
predictions_loss <- predict(loess_loss, newdata = nd, se = TRUE)
predictions_loss <- cbind(nd, predictions_loss)

ggplot() +
  geom_hline(yintercept = 0, linewidth = 1, color = "gold") +
  geom_line(aes(x = Low_Freq, y = snr_db), color = "purple", data = stimtrial, linewidth = 0.5) +
  geom_line(aes(x = Low_Freq, y = fit), color = "purple", linewidth = 1, data = predictions_loss) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - se.fit, ymax = fit + se.fit), fill = "purple", alpha = 0.5, data = predictions_loss) +
  xlab("Frequency (Hz)") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  scale_x_continuous(limits = c(20, 2100), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-75, 15), breaks = c(seq(-70, 10, 10)), expand = c(0,0)) +
  my_theme
```

# Vibration Transmission Analysis

## Short-Range Transmission

```{r short prep}
# choose between method 1 (vibration_transmission) and method 2 (vibration_transmission2)
# method 1 averages across spiders using logarithmic units 
# method 2 averages across spiders using linear units
method = vibration_transmission2
close_ind1 <- method %>% 
  filter(Distance == "close") %>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"),
         Site = factor(Site), 
         Site = fct_relevel(Site, "Rural", "Urban"),
         SpiderID = factor(SpiderID),
         Low_Freq = as.numeric(Low_Freq),
         Site_Treatment = factor(Site_Treatment),
         Site_Treatment= fct_recode(Site_Treatment, "Rural/Quiet" = "Rural_Quiet", "Rural/Loud" = "Rural_Loud", "Urban/Quiet" = "Urban_Quiet", "Urban/Loud" = "Urban_Loud"),
         Site_Treatment = fct_relevel(Site_Treatment, "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))

# go one step further and average the 3 trials at 3 positions for 15 spiders across site x treatment
close_ind <- method %>% 
  filter(Distance == "close") %>% 
  group_by(Low_Freq, Site, treatment, Site_Treatment) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db))%>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"),
         Site_Treatment = factor(Site_Treatment),
         Site_Treatment= fct_recode(Site_Treatment, "Rural/Quiet" = "Rural_Quiet", "Rural/Loud" = "Rural_Loud", "Urban/Quiet" = "Urban_Quiet", "Urban/Loud" = "Urban_Loud"),
         Site_Treatment = fct_relevel(Site_Treatment, "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))
```

```{r short loess}
# get the loess curve of urban loud from the average
close_ind1_ul <- close_ind1 %>% 
  filter(Site == "Urban",
         treatment == "Loud")
loess_close_ul <- loess(spider_snr_db ~ Low_Freq, data = close_ind1_ul, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Urban",
                  treatment = "Loud")
predictions_ul <- predict(loess_close_ul, newdata = nd, se = TRUE)
predictions_ul <- cbind(nd, predictions_ul)

# get the loess curve of rural loud from the average
close_ind1_rl <- close_ind1 %>% 
  filter(Site == "Rural",
         treatment == "Loud")
loess_close_rl <- loess(spider_snr_db ~ Low_Freq, data = close_ind1_rl, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Rural",
                  treatment = "Loud")
predictions_rl <- predict(loess_close_rl, newdata = nd, se = TRUE)
predictions_rl <- cbind(nd, predictions_rl)

# get the loess curve of urban quiet from the average
close_ind1_uq <- close_ind1 %>% 
  filter(Site == "Urban",
         treatment == "Quiet")
loess_close_uq <- loess(spider_snr_db ~ Low_Freq, data = close_ind1_uq, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Urban",
                  treatment = "Quiet")
predictions_uq <- predict(loess_close_uq, newdata = nd, se = TRUE)
predictions_uq <- cbind(nd, predictions_uq)

# get the loess curve of rural quiet from the average
close_ind1_rq <- close_ind1 %>% 
  filter(Site == "Rural",
         treatment == "Quiet")
loess_close_rq <- loess(spider_snr_db ~ Low_Freq, data = close_ind1_rq, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Rural",
                  treatment = "Quiet")
predictions_rq <- predict(loess_close_rq, newdata = nd, se = TRUE)
predictions_rq <- cbind(nd, predictions_rq)

predictions_close <- rbind(predictions_ul, predictions_uq, predictions_rl, predictions_rq) %>% 
  mutate(treatment = fct_relevel(factor(treatment), "Quiet", "Loud"))

ggplot() +
  geom_vline(xintercept = 1000, linetype = "dashed", color = "black") +
  geom_line(aes(x = Low_Freq, y = mean_snr, group = Site, color = Site), data = close_ind, alpha = 0.25) +
  geom_line(aes(x = Low_Freq, y = fit, group = Site, color = Site), data = predictions_close) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), group = Site, fill = Site), alpha = 0.5, data = predictions_close) +
  xlab("Frequency (Hz)") +
  ylab("Attenuation (∆ dB)") +
  scale_fill_manual("", values = two_colors) +
  scale_color_manual("", values = two_colors) +
  scale_x_continuous(limits = c(20, 2100), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
    scale_y_continuous(limits = c(-70, 15), breaks = c(seq(-60, 10, 10)), expand = c(0,0)) +
  my_theme +
  facet_wrap(~treatment, labeller = labeller(treatment = treatment.labs))
```

```{r short gam}
# generalize additive model
gam_model <- gam(spider_snr_db ~  Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = close_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = close_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Site_Treatment))

close_ind1_1000 <- close_ind1 %>% 
  filter(Low_Freq < 1000)
gam_model_1000 <- gam(spider_snr_db ~  Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = close_ind1_1000)
pairs(emmeans(gam_model_1000, ~Site_Treatment))

close_ind1_2000 <- close_ind1 %>% 
  filter(Low_Freq > 1000)
gam_model_2000 <- gam(spider_snr_db ~  Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = close_ind1_2000)
pairs(emmeans(gam_model_2000, ~Site_Treatment))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Site = levels(factor(close_ind1$Site)),
                             treatment = levels(factor(close_ind1$treatment))))
nd <- nd %>% 
  unite("Site_Treatment", Site, treatment, sep = "/", remove = FALSE)
predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_short <- cbind(nd, predictions) %>% 
  mutate(Site_Treatment = fct_relevel(factor(Site_Treatment), "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))

ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Site_Treatment), alpha = 0.15, data = close_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  #geom_vline(aes(xintercept = 300), linetype = "dashed") +
  #geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_short) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Site_Treatment), alpha = 0.5, data = predictions_short) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-63, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  scale_fill_manual("Site/Treatment", values = four_colors) +
  scale_color_manual("Site/Treatment", values = four_colors) +
  my_theme +
  #theme(legend.position = "none") +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
# 800 x 600
# 300 x 400
```

```{r short graphical abstract}
ga_short <- ggplot() +
  geom_vline(aes(xintercept = 300), linetype = "dashed") +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  annotate("rect", xmin = 300, xmax = 1000, ymin = -60, ymax = 0,
           alpha = .1,fill = "#d95f02") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_short) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Site_Treatment), alpha = 0.5, data = predictions_short) +
  scale_y_continuous(limits = c(-60, 0), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  scale_fill_manual("Site/Treatment", values = c("grey", "grey", "grey", "#d95f02")) +
  scale_color_manual("Site/Treatment", values = c("grey", "grey", "grey", "#d95f02")) +
  my_theme +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
```

```{r short temp graphs}
predictions_temp <- predictions_short %>% 
  filter(Site_Treatment == "Urban/Loud" | Site_Treatment == "Rural/Quiet")

ggplot() +
  #geom_line(aes(x = Low_Freq, y = mean_snr, color = Site_Treatment), alpha = 0.15, data = close_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  geom_vline(aes(xintercept = 300), linetype = "dashed") +
  geom_vline(aes(xintercept = 600), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_temp) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Site_Treatment), alpha = 0.5, data = predictions_temp) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-63, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  #scale_fill_manual("Site/Treatment", values = four_colors) +
  #scale_color_manual("Site/Treatment", values = four_colors) +
  my_theme +
  #theme(legend.position = "none") +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
```

```{r short assumptions}
# assumptions
test_close <- augment(gam_model, data = close_ind1)
resid_close <- ggplot(test_close, aes(x = .fitted, y = .resid)) + 
  geom_point() + 
  geom_smooth() +
  geom_hline(yintercept = 0) +
  xlab("Fitted Values") +
  ylab("Standardized \nResiduals") +
  my_theme

y <- quantile(test_close$.resid, c(0.25, 0.75))
x <- qnorm(c(0.25, 0.75))
slope <- diff(y)/diff(x)
int <- y[1L] - slope * x[1L]

qq_close <- ggplot(test_close, aes(sample = .resid)) + 
  stat_qq() + 
  geom_abline(slope = slope, intercept = int) +
  xlab("Theoretical Quantiles") +
  ylab("Sample Quantiles") +
  my_theme
ggarrange(resid_close, qq_close)
```

## Longer-Range Transmission

```{r longer prep}
far_ind1 <- method %>% 
  filter(Distance == "far") %>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"),
         Site = factor(Site), 
         Site = fct_relevel(Site, "Rural", "Urban"),
         Low_Freq = as.numeric(Low_Freq),
         Site_Treatment = factor(Site_Treatment),
         Site_Treatment= fct_recode(Site_Treatment, "Rural/Quiet" = "Rural_Quiet", "Rural/Loud" = "Rural_Loud", "Urban/Quiet" = "Urban_Quiet", "Urban/Loud" = "Urban_Loud"),
         Site_Treatment = fct_relevel(Site_Treatment, "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))

far_ind <- method %>% 
  filter(Distance == "far") %>% 
  group_by(Low_Freq, Site, treatment, Site_Treatment) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db))%>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"),
         Site = factor(Site),
         Site = fct_relevel(Site, "Rural", "Urban"),
         Site_Treatment = factor(Site_Treatment),
         Site_Treatment= fct_recode(Site_Treatment, "Rural/Quiet" = "Rural_Quiet", "Rural/Loud" = "Rural_Loud", "Urban/Quiet" = "Urban_Quiet", "Urban/Loud" = "Urban_Loud"),
         Site_Treatment = fct_relevel(Site_Treatment, "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))
```

```{r longer loess}
# get the loess curve of urban loud from the average
far_ind1_ul <- far_ind1 %>% 
  filter(Site == "Urban",
         treatment == "Loud")
loess_far_ul <- loess(spider_snr_db ~ Low_Freq, data = far_ind1_ul, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Urban",
                  treatment = "Loud")
predictions_far_ul <- predict(loess_far_ul, newdata = nd, se = TRUE)
predictions_far_ul <- cbind(nd, predictions_far_ul)

# get the loess curve of rural loud from the average
far_ind1_rl <- far_ind1 %>% 
  filter(Site == "Rural",
         treatment == "Loud")
loess_far_rl <- loess(spider_snr_db ~ Low_Freq, data = far_ind1_rl, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Rural",
                  treatment = "Loud")
predictions_far_rl <- predict(loess_far_rl, newdata = nd, se = TRUE)
predictions_far_rl <- cbind(nd, predictions_far_rl)

# get the loess curve of urban quiet from the average
far_ind1_uq <- far_ind1 %>% 
  filter(Site == "Urban",
         treatment == "Quiet")
loess_far_uq <- loess(spider_snr_db ~ Low_Freq, data = far_ind1_uq, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Urban",
                  treatment = "Quiet")
predictions_far_uq <- predict(loess_far_uq, newdata = nd, se = TRUE)
predictions_far_uq <- cbind(nd, predictions_far_uq)

# get the loess curve of rural quiet from the average
far_ind1_rq <- far_ind1 %>% 
  filter(Site == "Rural",
         treatment == "Quiet")
loess_far_rq <- loess(spider_snr_db ~ Low_Freq, data = far_ind1_rq, span = 0.5)
nd <- expand_grid(Low_Freq = seq(20, 2000, 5),
                  Site = "Rural",
                  treatment = "Quiet")
predictions_far_rq <- predict(loess_far_rq, newdata = nd, se = TRUE)
predictions_far_rq <- cbind(nd, predictions_far_rq)

predictions_far <- rbind(predictions_far_ul, predictions_far_uq, predictions_far_rl, predictions_far_rq) %>% 
  mutate(treatment = fct_relevel(factor(treatment), "Quiet", "Loud"),
         Site = fct_relevel(factor(Site), "Rural", "Urban"))

ggplot() +
  geom_vline(xintercept = 1000, linetype = "dashed", color = "black") +
  geom_line(aes(x = Low_Freq, y = mean_snr, group = Site, color = Site), data = far_ind, alpha = 0.25) +
  geom_line(aes(x = Low_Freq, y = fit, group = Site, color = Site), data = predictions_far) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - se.fit, ymax = fit + se.fit, group = Site, fill = Site), alpha = 0.5, data = predictions_far) +
  xlab("Frequency (Hz)") +
  ylab("Attenuation (∆ dB)") +
  scale_fill_manual("", values = two_colors) +
  scale_color_manual("", values = two_colors) +
  scale_y_continuous(limits = c(-72, 15), breaks = c(seq(-60,10, 10)), expand = c(0,0)) +
  scale_x_continuous(limits = c(20, 2100), breaks = c(20, 500, 1000, 1500, 2000), expand = c(0,0)) +
  my_theme +
  facet_wrap(~treatment, labeller = labeller(treatment = treatment.labs)) 
```

```{r longer gam}
# generalized additive model
gam_model <- gam(spider_snr_db ~ Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = far_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = far_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Site_Treatment))

far_ind1_1000 <- far_ind1 %>% 
  filter(Low_Freq < 1000)
gam_model_1000 <- gam(spider_snr_db ~  Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = far_ind1_1000)
pairs(emmeans(gam_model_1000, ~Site_Treatment))

far_ind1_2000 <- far_ind1 %>% 
  filter(Low_Freq > 1000)
gam_model_2000 <- gam(spider_snr_db ~  Site_Treatment + s(Low_Freq, by = Site_Treatment, k = 10), data = far_ind1_2000)
pairs(emmeans(gam_model_2000, ~Site_Treatment))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Site = levels(factor(far_ind1$Site)),
                             treatment = levels(factor(far_ind1$treatment))))
nd <- nd %>% 
  unite("Site_Treatment", Site, treatment, sep = "/", remove = FALSE)
predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_long <- cbind(nd, predictions) %>% 
  mutate(Site_Treatment = fct_relevel(factor(Site_Treatment), "Rural/Quiet", "Rural/Loud", "Urban/Quiet", "Urban/Loud"))

ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Site_Treatment), alpha = 0.15, data = far_ind) +
  #geom_vline(aes(xintercept = 350), linetype = "dashed") +
  #geom_vline(aes(xintercept = 600), linetype = "dashed") +
  #geom_vline(aes(xintercept = 100), linetype = "dashed") +
  #geom_vline(aes(xintercept = 250), linetype = "dashed") +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_long) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Site_Treatment), alpha = 0.5, data = predictions_long) +
  #scale_y_continuous(limits = c(-72, -33), breaks = c(seq(-70,-35, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 15), breaks = c(seq(-50,10,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 15), breaks = c(seq(-60, 10, 10)), expand = c(0,0)) + scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  scale_fill_manual("Site/Treatment", values = four_colors) +
  scale_color_manual("Site/Treatment", values = four_colors) +
  scale_linetype_manual("Site_Treatment", values = c("dashed", "solid", "dashed", "solid")) +
  my_theme +
  #theme(legend.position = "top") +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
```

```{r longer graphical abstract}
ga_longer <-ggplot() +
  geom_vline(aes(xintercept = 350), linetype = "dashed") +
  geom_vline(aes(xintercept = 600), linetype = "dashed") +
  annotate("rect", xmin = 350, xmax = 600, ymin = -60, ymax = 0,
           alpha = .1,fill = "#1b9e77") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_long) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - 1.96 * se.fit, ymax = fit + 1.96 * se.fit, fill = Site_Treatment), alpha = 0.5, data = predictions_long) +
  scale_y_continuous(limits = c(-60, 0), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  scale_fill_manual("Site/Treatment", values = c("grey", "#1b9e77", "grey", "grey")) +
  scale_color_manual("Site/Treatment", values = c("grey", "#1b9e77", "grey", "grey")) +
  scale_linetype_manual("Site_Treatment", values = c("dashed", "solid", "dashed", "solid")) +
  my_theme +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
```

```{r graphical abstract}
ggarrange(ga_short, ga_longer, ncol = 2)
#1200x500
```

```{r longer temp graphs}
predictions_temp <- predictions_long %>% 
  filter(Site_Treatment == "Urban/Loud" | Site_Treatment == "Rural/Quiet")

ggplot() +
  #geom_line(aes(x = Low_Freq, y = mean_snr, color = Site_Treatment), alpha = 0.15, data = close_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  geom_vline(aes(xintercept = 1450), linetype = "dashed") +
  geom_vline(aes(xintercept = 1650), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Site_Treatment), linewidth = 1, data = predictions_temp) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Site_Treatment), alpha = 0.5, data = predictions_temp) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 15), breaks = c(seq(-60, 10, 10)), expand = c(0,0)) + scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("Energy Loss through \nWeb Transmission (∆ dB)") +
  #scale_fill_manual("Site/Treatment", values = four_colors) +
  #scale_color_manual("Site/Treatment", values = four_colors) +
  my_theme +
  #theme(legend.position = "none") +
  guides(color = guide_legend(nrow = 2, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 2)) 
```

```{r longer assumptions}
# assumptions
test_close <- augment(gam_model, data = far_ind1)
resid_close <- ggplot(test_close, aes(x = .fitted, y = .resid)) + 
  geom_point() + 
  geom_smooth() +
  geom_hline(yintercept = 0) +
  xlab("Fitted Values") +
  ylab("Standardized \nResiduals") +
  my_theme

y <- quantile(test_close$.resid, c(0.25, 0.75))
x <- qnorm(c(0.25, 0.75))
slope <- diff(y)/diff(x)
int <- y[1L] - slope * x[1L]

qq_close <- ggplot(test_close, aes(sample = .resid)) + 
  stat_qq() + 
  geom_abline(slope = slope, intercept = int) +
  xlab("Theoretical Quantiles") +
  ylab("Sample Quantiles") +
  my_theme

ggarrange(resid_close, qq_close)
```

## Short vs Long Range by Site/Treatment

Use size = 18

```{r rq prep}
rq_ind1 <- method %>% 
  filter(Site_Treatment == "Rural_Quiet") %>% 
  mutate(SpiderID = factor(SpiderID),
         Low_Freq = as.numeric(Low_Freq),
         Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))

# go one step further and average the 3 trials at 3 positions for 15 spiders across site x treatment
rq_ind <- method %>% 
  filter(Site_Treatment == "Rural_Quiet") %>% 
  group_by(Low_Freq, Distance) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db)) %>% 
  mutate(Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))
```

```{r rq gam}
# generalize additive model
gam_model <- gam(spider_snr_db ~  Distance + s(Low_Freq, by = Distance, k = 10), data = rq_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = rq_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Distance))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Distance = levels(factor(rq_ind1$Distance))))

predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_rq <- cbind(nd, predictions)

rq <- ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Distance), alpha = 0.15, data = rq_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  #geom_vline(aes(xintercept = 300), linetype = "dashed") +
  #geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Distance), linewidth = 1, data = predictions_rq) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Distance), alpha = 0.5, data = predictions_rq) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("") +
  scale_fill_manual("Transmission Distance", values = c("black", "grey70")) +
  scale_color_manual("Transmission Distance", values = c("black", "grey70")) +
  my_theme +
  #theme(legend.position = "none") +
  ggtitle("Rural/Quiet") +
  guides(color = guide_legend(nrow = 1, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 1)) 
# 800 x 600
# 300 x 400
```

```{r rl prep}
rl_ind1 <- method %>% 
  filter(Site_Treatment == "Rural_Loud") %>% 
  mutate(SpiderID = factor(SpiderID),
         Low_Freq = as.numeric(Low_Freq),
         Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))

# go one step further and average the 3 trials at 3 positions for 15 spiders across site x treatment
rl_ind <- method %>% 
  filter(Site_Treatment == "Rural_Loud") %>% 
  group_by(Low_Freq, Distance) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db)) %>% 
  mutate(Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))
```

```{r rl gam}
# generalize additive model
gam_model <- gam(spider_snr_db ~  Distance + s(Low_Freq, by = Distance, k = 10), data = rl_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = rl_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Distance))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Distance = levels(factor(rl_ind1$Distance))))

predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_rl <- cbind(nd, predictions)

rl <- ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Distance), alpha = 0.15, data = rl_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  #geom_vline(aes(xintercept = 300), linetype = "dashed") +
  #geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Distance), linewidth = 1, data = predictions_rl) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Distance), alpha = 0.5, data = predictions_rl) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("") +
  scale_fill_manual("Transmission Distance", values = c("black", "grey70")) +
  scale_color_manual("Transmission Distance", values = c("black", "grey70")) +
  my_theme +
  #theme(legend.position = "none") + 
  ggtitle("Rural/Loud") +
  guides(color = guide_legend(nrow = 1, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 1)) 
# 800 x 600
# 300 x 400
```

```{r uq prep}
uq_ind1 <- method %>% 
  filter(Site_Treatment == "Urban_Quiet") %>% 
  mutate(SpiderID = factor(SpiderID),
         Low_Freq = as.numeric(Low_Freq),
         Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))

# go one step further and average the 3 trials at 3 positions for 15 spiders across site x treatment
uq_ind <- method %>% 
  filter(Site_Treatment == "Urban_Quiet") %>% 
  group_by(Low_Freq, Distance) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db)) %>% 
  mutate(Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))
```

```{r uq gam}
# generalize additive model
gam_model <- gam(spider_snr_db ~  Distance + s(Low_Freq, by = Distance, k = 10), data = uq_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = uq_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Distance))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Distance = levels(factor(uq_ind1$Distance))))

predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_uq <- cbind(nd, predictions)

uq <- ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Distance), alpha = 0.15, data = uq_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  #geom_vline(aes(xintercept = 300), linetype = "dashed") +
  #geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Distance), linewidth = 1, data = predictions_uq) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Distance), alpha = 0.5, data = predictions_uq) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("") +
  scale_fill_manual("Transmission Distance", values = c("black", "grey70")) +
  scale_color_manual("Transmission Distance", values = c("black", "grey70")) +
  my_theme +
  #theme(legend.position = "none") +
  ggtitle("Urban/Quiet") +
  guides(color = guide_legend(nrow = 1, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 1)) 
# 800 x 600
# 300 x 400
```

```{r ul prep}
ul_ind1 <- method %>% 
  filter(Site_Treatment == "Urban_Loud") %>% 
  mutate(SpiderID = factor(SpiderID),
         Low_Freq = as.numeric(Low_Freq),
         Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))

# go one step further and average the 3 trials at 3 positions for 15 spiders across site x treatment
ul_ind <- method %>% 
  filter(Site_Treatment == "Urban_Loud") %>% 
  group_by(Low_Freq, Distance) %>% 
  summarize(mean_snr = mean(spider_snr_db),
            se_snr = plotrix::std.error(spider_snr_db)) %>% 
  mutate(Distance = factor(Distance),
         Distance = fct_recode(Distance, "Longer-Range" = "far", "Short-Range" = "close"),
         Distance = fct_relevel(Distance, "Short-Range", "Longer-Range"))
```

```{r ul gam}
# generalize additive model
gam_model <- gam(spider_snr_db ~  Distance + s(Low_Freq, by = Distance, k = 10), data = ul_ind1)
gam_model_null <- gam(spider_snr_db ~ 1, data = ul_ind1)
anova(gam_model_null, gam_model, test = "F")
c(AIC(gam_model), AIC(gam_model_null))
sum1 <- summary(gam_model)
sum1
sum1$pTerms.table
gam.check(gam_model)
pairs(emmeans(gam_model, ~Distance))

nd <- data.frame(expand.grid(Low_Freq = seq(0, 2000, 5),
                             Distance = levels(factor(ul_ind1$Distance))))

predictions <- data.frame(predict(gam_model, newdata = nd, type = "response", 
                       se.fit = TRUE))
predictions_ul <- cbind(nd, predictions)

ul <- ggplot() +
  geom_line(aes(x = Low_Freq, y = mean_snr, color = Distance), alpha = 0.15, data = ul_ind) +
  geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_hline(aes(yintercept = 0), linetype = "solid") +
  #geom_vline(aes(xintercept = 300), linetype = "dashed") +
  #geom_vline(aes(xintercept = 1000), linetype = "dashed") +
  geom_line(aes(x = Low_Freq, y = fit, color = Distance), linewidth = 1, data = predictions_ul) +
  geom_ribbon(aes(x = Low_Freq, ymin = fit - (1.96 * se.fit), ymax = fit + (1.96 * se.fit), fill = Distance), alpha = 0.5, data = predictions_ul) +
  #scale_y_continuous(limits = c(-70, -43), breaks = c(seq(-70,-45, 5)), expand = c(0,0)) + scale_x_continuous(limits = c(900, 2100), breaks = c(seq(1000,2000,250)), expand = c(0,0)) +
  #scale_y_continuous(limits = c(-60, 3), breaks = c(seq(-50,0,10)),expand = c(0,0)) + scale_x_continuous(limits = c(0, 1100), breaks = c(0, 250, 500, 750, 1000), expand = c(0,0)) +
  scale_y_continuous(limits = c(-72, 3), breaks = c(seq(-60, 0, 10)), expand = c(0,0)) +   scale_x_continuous(limits = c(0, 2100), breaks = c(0, 500, 1000, 1500, 2000), expand = c(0,0)) +
  xlab("Frequency (Hz)") +
  #ylab("Energy Loss (∆ dB)\n\n") +
  ylab("") +
  scale_fill_manual("Transmission Distance", values = c("black", "grey70")) +
  scale_color_manual("Transmission Distance", values = c("black", "grey70")) +
  my_theme +
  #theme(legend.position = "none") +
  ggtitle("Urban/Loud") +
  guides(color = guide_legend(nrow = 1, override.aes = list(linewidth = 2))) +
  guides(fill = guide_legend(nrow = 1)) 
# 800 x 600
# 300 x 400
```

```{r combined plot}
require(grid)

figure <- ggarrange(rq, rl, uq, ul, nrow = 2, ncol = 2, common.legend = TRUE)

annotate_figure(figure, left = textGrob("Energy Loss through Web Transmission (∆ dB)", rot = 90, vjust = 1, gp = gpar(cex = 1.3, fontsize = size)))

#1000 x 800
```

# Silk Analysis

## Silk Mass 

```{r silk mass prep}
silk_sum <- silk %>% 
  group_by(Site, treatment) %>% 
  summarize(mean_silk = mean(dry_microg_body_mg), 
            se_silk = plotrix::std.error(dry_microg_body_mg))
```

```{r silk mass raw data}
silk_sum %>% 
  ggplot() +
  geom_point(aes(x = treatment, y = mean_silk, group = Site, color = Site)) +
  geom_errorbar(aes(x = treatment, ymin = mean_silk - se_silk, ymax = mean_silk + se_silk, group = Site, color = Site), width = 0.25) +
  xlab("Treatment") +
  ylab("Dry Silk Mass (µg) Corrected by Body Mass (mg)") +
  scale_fill_manual(values = c("#1b9e77", "#d95f02"),
                     labels = c("Rural", "Urban")) +
  scale_color_manual(values = c("#1b9e77", "#d95f02"),
                     labels = c("Rural", "Urban")) +
  theme_classic() +
  theme(axis.text = element_text(size = 14, color = "black"),
        axis.title = element_text(size = 14, color = "black"),
        legend.text = element_text(size = 14, color = "black"),
        legend.title = element_text(size = 14, color = "black"),
        legend.position = "top")
```

```{r silk mass stats}
shapiro.test(sqrt(silk$dry_microg_body_mg))
silk_model <- lm(sqrt(dry_microg_body_mg) ~ Site * treatment, data = silk)
car::Anova(silk_model)
#emmeans(silk_model, pairwise ~ treatment,  by = "Site")$contrasts
#emmeans(silk_model, pairwise ~ Site,  by = "treatment")$contrasts

# not normal
silk_loud <- silk %>% 
  filter(treatment == "Loud")
kruskal.test(dry_microg_body_mg ~ Site, data = silk_loud)
silk_quiet <- silk %>% 
  filter(treatment == "Quiet")
kruskal.test(dry_microg_body_mg ~ Site, data = silk_quiet)
```

```{r silk mass assumptions}
test_silk <- augment(silk_model, data = silk)
resid_silk <- ggplot(test_silk, aes(x = .fitted, y = .resid)) + 
  geom_point() + 
  geom_smooth() +
  geom_hline(yintercept = 0) +
  xlab("Fitted Values") +
  ylab("Standardized \nResiduals") +
  theme_classic() +
  theme(text = element_text(size = 14, color = "black")) +
  theme(axis.text.x=element_text(color="black", size=14)) + 
  theme(axis.text.y=element_text(color="black", size=14))

y <- quantile(test_silk$.resid, c(0.25, 0.75))
x <- qnorm(c(0.25, 0.75))
slope <- diff(y)/diff(x)
int <- y[1L] - slope * x[1L]

qq_silk <- ggplot(test_silk, aes(sample = .resid)) + 
  stat_qq() + 
  geom_abline(slope = slope, intercept = int) +
  xlab("Theoretical Quantiles") +
  ylab("Sample Quantiles") +
  theme_classic() +
  theme(text = element_text(size = 14, color = "black")) +
  theme(axis.text.x=element_text(color="black", size=14)) + 
  theme(axis.text.y=element_text(color="black", size=14)) 

ggarrange(resid_silk, qq_silk)
```

```{r silk mass graph}
nd = data.frame(expand_grid(Site = levels(factor(silk$Site)),
                treatment = levels(factor(silk$treatment))))

nd$fit <- predict(silk_model, newdata = nd, se.fit = TRUE, re.form = NA, type = "response")$fit
nd$SE <- predict(silk_model, newdata = nd, se.fit = TRUE, re.form = NA, type = "response")$se.fit
nd <- nd %>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"))

ggplot() +
  geom_jitter(aes(x = treatment, y = dry_microg_body_mg, group = Site, color = Site), size = 1, alpha = 0.5, width = 0.1, data = silk) +
  geom_line(aes(x = treatment, y = fit^2, group = Site, color = Site), position = position_dodge(width = 0.25), size = 1, data = nd) +
  geom_point(aes(x = treatment, y = fit^2, group = Site, color = Site), position = position_dodge(width = 0.25), size = 2.5, data = nd) +
  geom_errorbar(aes(x = treatment, ymin = (fit - 1.96 * SE)^2, ymax = (fit + 1.96 * SE)^2, group = Site, color = Site), width = 0.25, position = position_dodge(width = 0.25), size = 1, data = nd) +
  xlab("Treatment (4 Days)") +
  ylab("Dry Silk Mass (µg) \nCorrected by Body Mass (mg)") +
  scale_color_manual("", values = c("#1b9e77", "#d95f02")) +
  scale_y_continuous(limits = c(0, 26), breaks = c(0, 5, 10, 15, 20, 25), expand = c(0,0)) +
  my_theme
```

## Silk Moisture

```{r silk moisture prep}
moist_sum <- silk %>% 
  group_by(Site, treatment) %>% 
  summarize(mean_moist = mean(moisture_per), 
            se_moist = plotrix::std.error(moisture_per)) %>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"))
```

```{r silk moisture raw data}
moist_sum %>% 
  ggplot() +
  geom_point(aes(x = treatment, y = mean_moist, group = Site, color = Site)) +
  geom_errorbar(aes(x = treatment, ymin = mean_moist - se_moist, ymax = mean_moist + se_moist, group = Site, color = Site), width = 0.25) +
  xlab("Treatment") +
  ylab("Percent Silk Moisture") +
  scale_fill_manual(values = c("#1b9e77", "#d95f02")) +
  scale_color_manual(values = c("#1b9e77", "#d95f02")) +
  theme_classic() +
  theme(axis.text = element_text(size = 12, color = "black"),
        axis.title = element_text(size = 12, color = "black"),
        legend.text = element_text(size = 12, color = "black"),
        legend.title = element_text(size = 12, color = "black"),
        legend.position = "top")
```

```{r sik moisture stats}
shapiro.test(sqrt(silk$moisture_per))
moist_model <- lm(sqrt(moisture_per) ~ Site * treatment, data = silk)
car::Anova(moist_model)

#moist_model <- betareg(moisture_per ~ Site * treatment, data = silk)
#car::Anova(moist_model)

# not normal
silk_urban <- silk %>% 
  filter(Site == "Urban")
kruskal.test(moisture_per ~ treatment, data = silk_urban)
silk_rural <- silk %>% 
  filter(Site == "Rural")
kruskal.test(moisture_per ~ treatment, data = silk_rural)

emmeans(moist_model, pairwise ~ treatment,  by = "Site")$contrasts
emmeans(moist_model, pairwise ~ Site,  by = "treatment")$contrasts
```

```{r silk moisture graph}
# predictions
nd = data.frame(expand_grid(Site = levels(factor(silk$Site)),
                treatment = levels(factor(silk$treatment))))

nd$fit <- predict(moist_model, newdata = nd, se.fit = TRUE, re.form = NA, type = "response")$fit
nd$se <- predict(moist_model, newdata = nd, se.fit = TRUE, re.form = NA, type = "response")$se.fit
nd <- nd %>% 
  mutate(treatment = factor(treatment),
         treatment = fct_relevel(treatment, "Quiet", "Loud"))

ggplot() +
  geom_jitter(aes(x = treatment, y = moisture_per, group = Site, color = Site), size = 1, alpha = 0.5, width = 0.1, data = silk) +
  geom_line(aes(x = treatment, y = fit^2, group = Site, color = Site), position = position_dodge(width = 0.25), size = 1, data = nd) +
  geom_point(aes(x = treatment, y = fit^2, group = Site, color = Site), position = position_dodge(width = 0.25), size = 2.5, data = nd) +
  geom_errorbar(aes(x = treatment, ymin = (fit - 1.96 * se)^2, ymax = (fit + 1.96 * se)^2, group = Site, color = Site), width = 0.25, position = position_dodge(width = 0.25), size = 1, data = nd) +
  xlab("Treatment (4 Days)") +
  ylab("Proportion of Silk Moisture") +
  scale_color_manual("", values = c("#1b9e77", "#d95f02")) +
  scale_y_continuous(limits = c(0, 0.6), breaks = c(0, 0.2, 0.4, 0.6), expand = c(0,0)) +
  my_theme

```

## Silk Table of Stats

```{r silk table}
silk_stats <- read.csv(file = "data/silk_table.csv", header = TRUE) %>% 
  mutate(dependent = factor(dependent),
         dependent = fct_recode(dependent, "Dry Silk Mass (µg) per Body Mass (mg)" = "silk_mass", 
                            "Prop. Silk Moisture" = "silk_moist"),
         independent = factor(independent),
         independent = fct_recode(independent, "Collection Site" = "Site",
                               "Treatment" = "Treatment",
                               "Collection Site x Treatment" = "Site_Treatment"),
         df = factor(df),
         df = fct_recode(df, " 56" = "56", "56" = "_56"),
         p_value = factor(p_value),
         p_value = fct_recode(p_value, "   0.117" = "0.117",
                        "   0.777" = "0.777",
                        "   0.124" = "0.124",
                        "   0.463" = "0.463",
                        "   0.089 ." = "0.089",
                        "   0.157" = "0.157")) 

colnames(silk_stats) <- c('Dependent Variable', 'Residual DF', 'Independent Variable', 'F-Value', 'P-Value')

flextable(silk_stats[, 1:5]) %>% 
  autofit() %>% 
  merge_v(j = c('Dependent Variable', 'Residual DF'), part = "body") %>% 
  theme_vanilla() %>% 
  hline(i = c("1", "2", "4", "5"), 
        j = c("Independent Variable", "F-Value", "P-Value"), 
        part = "body", 
        border = officer::fp_border(width = 0, color = "black")) %>% 
  bold(bold = TRUE, part = "header") %>% 
  fontsize(size = 10, part = "all") %>% 
  align(align = "left", part = "all") %>%
  valign(valign = "top", part = "all")  #%>% 
  #save_as_image(path = here::here("silk_table.png"))
```
